requirements.txt
fastapi==0.115.3
httpx==0.27.2
python-dotenv==1.0.1
uvicorn==0.32.0
sqlite3-binary
